requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==4.2.16
django-crispy-forms==2.3
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2